refactor: move types.ts
