+++ packages/core/helper-plugin/src/components/AnErrorOccurred.tsx
@@ -1,10 +1,8 @@
-import React from 'react';
-
 import { EmptyStateLayout, EmptyStateLayoutProps } from '@strapi/design-system';
 import { ExclamationMarkCircle } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 
-import type { TranslationMessage } from '../../types';
+import type { TranslationMessage } from '../types';
 
 export type AnErrorOccurredProps = Omit<EmptyStateLayoutProps, 'content' | 'icon'> & {
   content?: TranslationMessage;
